1. Убрана бизнес логика из классов контроллеров а также из репозиториев, все перенесено в сервис 2. Фикс ошибки ValidationException теперь возвращается с правильным кодом 3. Тесты разделены на тесты для фильмов и тесты для пользователей

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,25 +1,37 @@
 package ru.yandex.practicum.filmorate.service;
 
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
+import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.FilmStorage;
 
+import java.time.LocalDate;
 import java.util.Collection;
 import java.util.List;
 
 @Service
+@Slf4j
 public class FilmService {
 
     private final FilmStorage filmStorage;
+    private final UserService userService;
+    private static final LocalDate MIN_RELEASE_DATE = LocalDate.of(1895, 12, 28);
 
     @Autowired
-    public FilmService(FilmStorage filmStorage) {
+    public FilmService(FilmStorage filmStorage, UserService userService) {
+        this.userService = userService;
         this.filmStorage = filmStorage;
     }
 
     public Film getFilmByIdFromStorage(Long id) {
-        return filmStorage.getFilmById(id);
+        Film film = filmStorage.getFilmById(id);
+        if (film == null)
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        return film;
     }
 
     public boolean containsKey(Long id) {
@@ -30,23 +42,45 @@ public class FilmService {
         return filmStorage.getPopularFilmLimitCount(count);
     }
 
-    public Film addFilmInStorage(Long id, Film film) {
-        return filmStorage.addFilm(id, film);
+    public Film addFilmInStorage(Film film) {
+        if (film.getReleaseDate().isBefore(MIN_RELEASE_DATE)) {
+            log.error("Дата релиза {} раньше допустимой", film.getReleaseDate());
+            throw new ValidationException("Дата релиза должна быть не раньше 28 декабря 1895 года");
+        }
+        film.setId(filmStorage.getNextId());
+        return filmStorage.addFilm(film);
     }
 
     public Film deleteLikeFromFilmInStorage(Long id, Long whoIsDeleted) {
+        if (!userService.containsKey(whoIsDeleted) || !this.getFilmByIdFromStorage(id).getIdOfUsersWhoLikedThisFilm().contains(whoIsDeleted))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
         return filmStorage.deleteLikeFromFilm(id, whoIsDeleted);
+    }
+
+    public Film postLikeOnFilmInStorage(Long id, Long userId) {
+        if (!this.containsKey(id) || !userService.containsKey(userId))
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND);
+        this.getFilmByIdFromStorage(id).getIdOfUsersWhoLikedThisFilm().add(userId);
+        return this.getFilmByIdFromStorage(id);
     }
 
     public Collection<Film> getAllFilmsFromStorage() {
         return filmStorage.getAllValues();
     }
 
-    public long getNextIdFromStorage() {
-        return filmStorage.getNextId();
-    }
-
     public Film updateFilmInStorage(Film film) {
+        if (film.getId() == null) {
+            log.error("Ошибка валидации порядкового номера(id) фильма");
+            throw new ValidationException("Id должен быть указан");
+        }
+        if (!this.containsKey(film.getId())) {
+            log.error("Ошибка - фильм не найден");
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Пост с id = " + film.getId() + " не найден");
+        }
+        if (film.getReleaseDate().isBefore(MIN_RELEASE_DATE)) {
+            log.error("Дата релиза {} раньше допустимой", film.getReleaseDate());
+            throw new ValidationException("Дата релиза должна быть не раньше 28 декабря 1895 года");
+        }
         return filmStorage.updateFilm(film);
     }
 }
